trpc main composer finished, jsut needs cleaning up
