+++ main_process/main_httpController.js
@@ -305,7 +305,7 @@ const httpController = {
       headersResponse[':status'] = response.status;
       const receivedCookie = headersResponse['set-cookie'];
       headersResponse.cookies = receivedCookie;
-      console.log(receivedCookie);
+
       const contentType = response.headers.get('content-type');
       const contents = /json/.test(contentType)
         ? await response.json()
